独立查询历史记录的接口

+++ src/main/java/com/nutrition/nutritionservice/dao/UserHistoricalCuisineDao.java
@@ -13,7 +13,7 @@ public interface UserHistoricalCuisineDao {
 
     UserHistoricalCuisineVo selectLastCuisineByUuid(String uuid);
 
-    long insert(UserHistoricalCuisineVo record);
+    int insert(UserHistoricalCuisineVo record);
 
     List<UserHistoricalCuisineVo> selectByUuid(String uuid);
